requirements.txt
# Common
isort
flake8
pdbpp
pytest
ipython
httpx
requests
uvicorn
fastapi
typer

# Database
#sqlalchemy
tortoise-orm